Rename to ptychography40; update links
Closes #37
