+++ src/ptychography40/stitching/stitching.py
