+++ tests/test_smoke.py
@@ -1,6 +1,6 @@
-import ptychography
-import ptychography.reconstruction.ssb
-import ptychography.stitching.stitching
+import ptychography40
+import ptychography40.reconstruction.ssb
+import ptychography40.stitching.stitching
 
 
 def test_stuff():
