+++ tox.ini
@@ -3,8 +3,8 @@ envlist = py{36,37}
 
 [testenv]
 commands=
-    pytest --durations=5 --cov=ptychography --cov-report=term --cov-report=html --cov-report=xml {posargs:tests/}
-    pytest --doctest-modules src/ptychography/
+    pytest --durations=5 --cov=ptychography40 --cov-report=term --cov-report=html --cov-report=xml {posargs:tests/}
+    pytest --doctest-modules src/ptychography40/
 deps=
     -rtest_requirements.txt
 setenv=
